refactor: Improve rental detail package structure (#67)
