+++ app/src/main/java/com/example/rentit/common/model/PriceSummaryUiModel.kt
@@ -1,0 +1,6 @@
+package com.example.rentit.common.model
+
+data class PriceSummaryUiModel(
+    val label: String,
+    val price: Int
+)
