+++ app/src/main/java/com/example/rentit/presentation/rentaldetail/RentalDetailRoute.kt
@@ -1,4 +1,4 @@
-package com.example.rentit.presentation.rentaldetail.renter
+package com.example.rentit.presentation.rentaldetail
 
 import android.os.Build
 import androidx.annotation.RequiresApi
@@ -8,12 +8,13 @@ import androidx.compose.runtime.getValue
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import androidx.navigation.NavHostController
+import com.example.rentit.presentation.rentaldetail.renter.RentalDetailRenterScreen
 
 @RequiresApi(Build.VERSION_CODES.O)
 @Composable
 fun RentalDetailRenterRoute(navHostController: NavHostController) {
 
-    val viewModel: RentalDetailRenterViewModel = hiltViewModel()
+    val viewModel: RentalDetailViewModel = hiltViewModel()
     val uiModel by viewModel.uiModel.collectAsStateWithLifecycle()
 
     val scrollState = rememberScrollState()
