--- app/src/main/java/com/example/rentit/presentation/rentaldetail/common/model/PriceItemUiModel.kt
@@ -1,6 +1,0 @@
-package com.example.rentit.presentation.rentaldetail.common.model
-
-data class PriceItemUiModel(
-    val label: String,
-    val price: Int
-)
